Training accounts need special treatment.
Currently training accounts are all "system" users so we don't
have to have an email address per account.  This causes a problem
as the Person.active.all() function skips them, so locking/unlocking
doesn't work for them.

Instead we use Person.objects.all() to access them and filter with
a regular expression.